'Diana DCS commit implementacion consultas por idbuyer, id compra y fechas de compra
