+++ Marketplace/purchase.backend/src/main/java/co/edu/uniandes/csw/miso4204/purchase/logic/ejb/PurchaseLogicService.java
@@ -30,6 +30,7 @@ Source generated by CrudMaker version 1.0.0.qualifier
 
 package co.edu.uniandes.csw.miso4204.purchase.logic.ejb;
 
+import co.edu.uniandes.csw.miso4204.purchase.logic.dto.PurchasePageDTO;
 import co.edu.uniandes.csw.miso4204.purchase.persistence.PurchasePersistence;
 
 public class PurchaseLogicService extends _PurchaseLogicService {
@@ -38,4 +39,11 @@ public class PurchaseLogicService extends _PurchaseLogicService {
 		persistance = new PurchasePersistence();
 	}
 	
+        public PurchasePageDTO getPurchasesSearch(Integer page, Integer maxRecords, String ini_date, String end_date, Long id_purchase){
+            return persistance.getPurchasesSearch(page, maxRecords, ini_date, end_date, id_purchase); 
+	}
+        
+        public PurchasePageDTO getPurchasesBuyer(Integer page, Integer maxRecords, Long id){
+            return persistance.getPurchasesBuyer(page, maxRecords, id); 
+	}
 }
